Modify resolveNew to operate differently for generic initializers
During resolution of generic initializers, we won't know the full instantiated
type for the "this" argument until we have resolved Phase 1 of the appropriate
initializer body.  Prior to this change, the determining of the appropriate
candidate function for the PRIM_NEW call would fail because it expected all the
actuals to be fully instantiated and the "this" actual would never be for a
generic AggregateType.  With this change, we convince resolution to allow us to
find the match when the "this" actual is still generic, and then update the
initializer during resolution of the body.

Details:

symbol.cpp:
  Change FnSymbol::hasGenericFormals to recognize when it is operating on an
initializer and to not complain about the "this" actual being generic in that
situation.

initializerRules.cpp:
  Modify buildClassAllocator and when it is called to respond to generic
initializers.  We don't want to create an _new method for a generic class, as we
will never use the allocator.  We only want an _new method for the generic
instantiations.  At normalize, we won't know the instantiations, so skip
creating the _new method if the class is generic.  Modify buildClassAllocator
itself to operate slightly differently when called from resolution: in this
case, don't force the function to find its init function through resolution as
we will have the fully resolved version at hand.  Also, don't set the
retExprType for the function (as this will cause the type constructor to be
called and generic instantiations don't have a type constructor); instead, set
the retType directly.

initializerResolution.h:
initializerResolution.cpp:
  Copy some strategies from functionResolution.cpp and generics.cpp, and use my
newly inserted helper functions for others, to resolve initializers.

I copied filterCandidate, gatherCandidates and friends and renamed the copy, as
they were pretty trivial and so not a problem to redo.  One or two bits that
were specific to functions/methods that aren't constructors were also removed in
the process.

I made a version of instantiateSignature specifically for initializers, which
doesn't complain when the "this" actual used for the instantiation is generic,
and which skips portions of the body that are definitely not relevant for
initializers (and in some cases would result in incorrect behavior).

modAndResolveInitCall handles what was previously in resolveNew for
initializers, and extends it in the case of generics.  For "_new" calls, it
inserts the type argument to allocate, and resolves the call.  For "init" calls,
it calls resolveInitializer and in the case of generics, updates the type of the
"this" actual to use the instantiation rather than the generic type.  It then
replaces the "init" call with a call to the newly created "_new" method if the
type is a class, and resolves the call and the function.

resolveInitializer is a very simplified version of resolveExpr's handling of
calls.  It calls out to resolveInitCall and resolveMatch.

resolveInitCall is very similar to resolveNormalCall, but utilizes my specific
initializer versions for gathering candidates and removes a lot of logic that is
unnecessary for initializers, such as ref/value pairs, matches against the name
being "=" (which won't be true ever), and logic associated with the modification
of const fields (which is permitted in Phase 1 but not in Phase 2).

resolveMatch is very similar to resolveFns, but does some initializer-specific
actions and removes unnecessary logic for extern functions, iterators, type
constructors and FLAG_PRIVATIZED_CLASS.  The initializer-specific actions are
centered around generics: it explicitly calls setNextGenericField; remembers
that the type was generic prior to resolving the initializer; and after
resolving the initializer, if it was on a generic class, builds the "_new"
function and normalizes it.

I also modified temporaryInitializerFixup to not modify the calls I insert in
modAndResolveInitCall, as that will result in incorrect behavior.

resolution.h:
functionResolution.cpp:

  Export resolveReturnType for use in my initializer-specific version of
resolveFns.

Modify ResolutionCandidate::computeSubstitutions to take in a boolean indicating
that the candidate function is an initializer.  This boolean is used when
creating the list of substitutions to avoid making one for the "this" argument,
as the combination of the "this" argument being generic until the completion of
Phase 1's resolution and the placement of the "this" argument so early in the
argument list means that we would otherwise end up in an infinite loop when
determining the substitutions should a later generic argument have a default
expression (because the "substitution" won't change whether the function is
considered generic, and the handling of default expressions for generic
arguments requires us to have a substitution for every prior generic argument in
case the default expression refers to it).  I debated about modifying
computeSubstitutions and its support function, computeGenericSubs, to determine
whether it should avoid creating a substitution for the "this" argument on its
own, but decided to wait to see if my reviewer also thought that was a good
idea.

Changed resolveNew to continue to resolve against "init" instead of "_new" when
the type is generic.  Moved and replaced the code which was special casing for
"init" methods to happen in initializerResolution.

Removed the quick fix for recursively calling initializers from resolveFns
proper, leaving it only in initializerResolution's version.