Avoid accidental clean up of my generic instantiations for initializers
Prior to this change, resolution was cleaning up every AggregateType whose
defaultTypeConstructor either did not exist or was not resolved.  This caused
problems because the generic instantiations I create for types with initializers
do not need or define the defaultTypeConstructor.  With this change, I add a
marker that is set when the initializer is resolved, so that the aforementioned
clean up will know to keep my generic instantiations around.