Minor set up fix for generic var/const fields
This won't fully implement support for them, but it will prevent an assertion
error I was seeing and instead allow them to reach the friendly error message.

I had forgotten that the type of a symbol will always be something instead of
NULL, so should have been comparing against dtUnknown instead of NULL when
determining if a field was a generic var or const.