Fix bug with reusing a generic initializer with the same generic arguments
The problem arose because I was getting the correct instantiation of the
function, but because the actual was still generic, canDispatch thought they
didn't match.  I modified canDispatch to check if the actual had a generic type
and the formal type was an instantiation of that generic type, and if so, assert
we're working with an initializer and return true.

This fixed the two test cases that were failing for me.