Add four tests for generic initializers with type fields
One verifies that the same type instantiation is used when reached through two
different initializers, one verifies that the same initializer instantiation is
used when reached through two identical calls, one verifies that the same
initializer instantiation is used when reached through two calls using the same
value for the generic argument but different values for the concrete argument,
and the last verifies that the same initializer is usable with two different
values for the generic argument (leading to two different instantiations).

The middle two failed prior to the last commit, but pass now.