Stop testing displayRepresentation by default
In my original take on this array API test, I tested
displayRepresentation for arrays even though the test was ostensibly
meant to test the user API and I consider displayRepresentation to be
a developer-oriented flag.

This resulted in regressions for this new test in numa testing whose
representation is different (duh).  It also required the tests to be
updated in the array view branch where they're different.  For those
reasons, I disabled that part of the test by default, though one can
throw a config param to re-enable it (though I haven't done that in
any of these tests for similar maintenance issues).