At Mike's suggestion, stop returning a DefExpr from my handling of initializers
The return type for modAndResolveInitCall was from an earlier iteration of the
change.  Mike pointed out that there was no functional need to call resolveExpr
on a DefExpr, and that there were places where I didn't even utilize the
variable.  This commit moves stops returning a DefExpr, stops calling
resolveExpr on it, and stops creating the "insertBefore" variable used for it
in the first place as an additional clean up.