Remove configuration-specific .good files now that leaks are to 0
With the array views work, I believe all of the leaks due to this
test have gone to 0 in all these configurations, so the standard
domainMaps.good file should now be good enough.  I also removed
the skipif on this basis.

Unfortunately, the .suppressif seems like it needs to stick around
because we still have leaks with baseline testing... :(