Add a variation of a test that fails for param fields
This test demonstrates that, though using the param field as the omitted
initialization for a concrete field causes problems, having the field's default
initialization utilize the param field does not impact the user's ability to
explicitly initialize the field during Phase 1.

This narrows down the source of the issue, slightly.