Lots of improvements to initial topology support.
The major change here is remove the runtime chpl_topo_showMemLocality()
function which printed locality information, and in place of that add
chpl_topo_getMemLocality() which returns it to the caller.  Then, the
array-loc test has been adjusted to do the locality interpretation and
printing itself.

The test now has a .prediff to select between .good files appropriate
for when we do and don't have libnuma support.  The old 1-1.good and
1-2.good files have been removed because the test now produces the same
result for the two .execopts entries.