Upgrade to jemalloc 4.5.0
jemalloc 4.5.0 includes a significant number of bug fixes, most of which are
the result of significantly improved CI testing. It fixes a number of bugs
introduced in the 4.X series. The one most important to us is:

 - "Restructure *CFLAGS configuration, so that CFLAGS behaves typically, and
   EXTRA_CFLAGS provides a way to specify e.g. -Werror during building, but not
   during configuration."

which will allow us to remove an osx+gnu workaround and simplify how we specify
CFLAGS to jemalloc.

For the full changelog see:
 - https://github.com/jemalloc/jemalloc/releases/4.5.0