Simplify our jemalloc makefile
With jemalloc 4.5.0, CFLAGS now behaves in a manner similar to our other
third-party libraries, which means we can use it instead of EXTRA_CFLAGS. This
allows us to cleanup and simplify our jemalloc makefile.

In the past using CFLAGS was not recommended because it prevented configure
from appending any flags it wanted to. Prior to 4.4.0 using EXTRA_CFLAGS was
the recommended option, but with 4.4.0 EXTRA_CFLAGS no longer got appended at
configure time, so flags specified in it didn't affect configure tests. For us,
this meant our optimization level wasn't set for configure tests, so jemalloc
was unable to detect that __builtin_unreachable was broken for osx+gnu (so we
had to workaround it.)

Now CFLAGS impact configure tests, and configure is able to tack on its own
CFLAGS. EXTRA_CFLAGS is only used for options that you want to set, but don't
want to impact configure tests (such as -Werror) so we no longer want to use
it.

This allows us to simplify our makefile for jemalloc and remove our lame
workaround for broken __builtin_unreachable on osx+gnu.

See https://github.com/jemalloc/jemalloc/pull/653,
https://github.com/jemalloc/jemalloc/issues/619, and
https://github.com/jemalloc/jemalloc/pull/504 for more info