Add test of zippering different views together
This test was motivated while I was doing code cleanup last night
and thinking about it more this morning -- I came away worried
that I'd gotten the 'these' iterators incorrect for array views,
but it turns out that they were correct.  The test is a nice
little one, though, so seemed worth keeping.