Do localization in the array allocator, not the memory it acquires.
With wider testing, it turned out that the first attempt to localize
arrays couldn't be made to work.  There were a couple of problems.  The
first was that if the array element type (eltType) was (Chapel) string,
the compiler insisted that the sizeof() function, which was a C extern,
should be applied to c_string instead.  The compiler's concern for the
possible mistake was appreciated, but unfortunately that isn't the same
thing.  Then when we tried to work around that, it turned out that
getting the sizeof() eltType was not always possible anyway.  (In
current code, wherever we need sizeof(something) to be passed to C code,
we do it inside the code generator as a side effect of expanding some
other primitive, such as "array_alloc".)

So here, we adjust our approach to do the localization call from inside
the runtime's array allocation functions.  The change is straightforward
enough: instead of passing the localization type and sublocale to the
runtime directly from ChapelBase:_ddata_allocate(), pass it to the
array_alloc primitive which the compiler turns into a call to the
runtime chpl_array_alloc() code.  There is also a small amount of
codegen:expr work to support the change to the array_alloc primitive,
but just in the number of arguments, nothing deep.  As a result, we can
use the sizeof(eltType) that the compiler produces as a side effect of
expanding the array_alloc primitive and don't need to code our own.

While here, I also moved the chpl(_wide)?_array_(alloc|free) functions
from chpl-comm-compiler-macros.h to chpl-mem.h, since they really had
never belonged in the former in the first place, and at the same time
renamed them as chpl_mem_(_wide)?_array_(alloc|free).