fix return array through arg cases
proc returnArrayArg(X) and proc returnArrayArgTyped(X:[1..size] real) are
meant to "return" the array by filling in the values of the argument.
But, they previously included a return statement. That return statement
was not needed.

Before the arrays rework, these functions wouldn't have a version made to
copy the argument out to the return variable (because of details of how
the implementation worked). But that does happen after the arrays rework.

Certainly we could consider re-enabling an optimization to avoid copies
from returning when a function is called but the return value is not
used. However that's unlikely to be important in real benchmarks.

When investigating performance, I also noticed a slight relative
slow-down in performance for the 1st thing tested here. So, I
added a warm-up loop that is not timed to try to tamp that down.