Remove '=>'-in-constructor error test
This test escaped my deprecation of '=>' commits earlier because I was
mistakenly grepping for tests that used '=>' and 'new' on the same
line and was mis-accounting for the number of tests I needed to
update.  Here I'm retiring it because we already have a placeholder
for the new deprecation warning and this test itself doesn't add
additional value.  Thanks to @noakesmichael for saving me the
embarrassment of having this fail by updating its .good earlier.