Don't localize if hugepages and/or NIC-registration will prevent it.
Based on past investigations, in all likelihood if we try to localize
hugepages or NIC-registered memory we'll fail.  Since we expect to get
locality in these situations in other ways in the future, for now turn
off setting the locality of already-allocated memory for configurations
comm=ugni, and comm=gasnet with seg!=everything.  (Note that these are
not configurations that we currently test nightly anyway, so this just
prevents people doing exploratory work from encountering problems in
them.)