Review feedback: the locality of a NULL pointer should be "any".
Based on review feedback, make chpl_topo_getMemLocality() return
sublocale "any" for a NULL pointer.  Granted this check could be done by
the caller, but we'd likely have to do it everywhere we didn't have a
strong notion of where the address had came from, so let's just do it in
the topology module itself.