Add two new tests, demonstrating the issue with using params in later fields
These tests, along with the similarly named use-param-in-concrete-default2.chpl,
exercise the cases that do and do not work for utilizing param fields in the
initialization of later fields.

use-param-in-concrete-default.chpl exercises the behavior when we rely on the
omitted initialization of the var field, and are thus inserting initialization
that uses the param field.  It fails during resolution, because of the previous
issue where we don't preface field accesses in our inserted initialization with
"this." (but in a different manner than using ordinary var/const fields, as
params are removed during this operation and rely on the presence of the "this."
to find them and clean them up).  Mike is working on this issue.

use-param-in-concrete-default3.chpl exercises using the param explicitly after
it has been initialized during Phase 1 of the user body.  This confirms that the
problem is solely with omitted initialization.