Add a workaround for an issue with RVF'ing a reduce intent shadow var
After PR #5134, the _build_tuple_always_allow_ref function used in the
implementation of reduce intents no longer caused remote value forwarding
to decide references were mutable. The result was forwarding and
localizing a reference to a reduction shadow variable in this test

  parallel/forall/vass/binary-tree-spawn.no-recurse.chpl

This PR adds a simple workaround to resolveIntents to restore something
like the previous behavior. However, it would be better for forall/task
intents to clearly mark the arguments in a way that distinguishes these
cases for references:

 * not const
 * const
 * later analysis should decide if it's const