Add three new tests of initializers on types with generic var fields
The first verifies that we get the same type when we pass in the exact same
argument to the initializer in multiple calls.  The second verifies that we
get the same type when we pass in two different arguments of the same type.  The
third verifies that we get two different types when we pass in two arguments of
different types.