Restore previous level of memory leaks in studies/amr code
In introducing a user constructor to this class, I introduced a second
call to the initialize() routine which resulted in a memory leak by
allocating the signature() classes twice.  I thought that this was
required for some reason when I introduced them, but the more obvious
fix is to just stop using the initialize() routine (which is also "a
good thing" (TM)).  The comment suggested that it had to be used since
the object was generic, but that doesn't hold up unless we had a bug
at the time that we were working around...