Remove remnants of the array indexing optimization
Remove a stop-gap optimization that eliminated the inner multiply in dsiAccess.
Happily, this optimization has been obsoleted by the array-views (#5338) work.
With array-views we're able to avoid the inner-multiply in more principled and
fine-grain manner (see #5338's PR message for more info.)

This just removes the remnants of the optimization that weren't already removed
with the array-views merge.