Do not insert an autoCopy for chpl__iterLF
ArrayViews exposed an issue in callDestructors where an autoCopy was
inserted for a chpl__iterLF variable with an ArrayView type. The
autoCopy results in a non-view type, leading to compilation issues
later on. Vass and I feel that this autoCopy should be unnecessary
given the usage of chpl__iterLF variables, so for now we will
test for FLAG_CHPL__ITERLF and avoid inserting an autoCopy.