Made the compile-time check contingent on --verify
David points out this. Since `maybeCalled()` check is heuristic,
it may return a false positive, in which case the compiler will
crash with INT_FATAL when there is no real problem.

While at the moment this is unlikely, it may become likely
as we increase the use of recursive iterators. Also, unrelated
changes in the compiler (such as the upcoming ForallStmt node,
which triggered the original bug) may trigger this scenario.

To avoid this risk, I change this check to run only with --verify.
This way it will happen in our nightly testing and will not happen
in user mode.