Add tests of initializers with default values for arguments when type is generic
Adds a test each for when the type is generic due to a type, param, or generic
var field, and the initializer defined for that type has an argument with a
default value.  Prior to the previous commit, these would all fail because we
insert the wrapper before we know the generic instantiation of the "this"
argument.