Leak CHPL_COMM_WORLD_REPLICATED(1) as temp solution
This isn't as bad as it sounds: The result is a leak size on the
order of numLocales at the end of MPI multilocale programs. Obviously,
a leak-free solution would be preferable in the long-term, but this
fixes the dereference nil bug for now.

Also, remove unecessary if/else block in MPI module

The coforall over locales will effectively do the same thing as the
else-block for numLocales == 1