Fix a bug Brad noticed with default array intent
Normalize included a check (from PR #307) that a ref var can't be assigned from
a const variable. Normalize was calling ArgSymbol::isConstant() which in turn
calls Type::isDefaultIntentConst().  But, Type::isDefaultIntentConst returned
`true` in this case when called on dtUnknown.

For arguments with ref-if-modified blank intent (such as arrays), the error
can't be discovered in normalize and instead it must be discovered after
resolution. checkResolved.cpp already includes code to find this error. I just
converted it from an internal error to a user-facing one.

I'm of half a mind to entirely remove the check in normalize.cpp normRefVar
emitting the error

   Cannot set a non-const reference to a const variable.

as I wouldn't generally expect normalize to be a reasonable place to do error
checking about const-ness.