Clean up, show issue
This commit includes some trivial renamings and also
some changes to populateForwardingMethods that were necessary
in an investigation of allowing forwarding methods to participate
in normal overload resolution (so as to change the precedence
of coercion vs delegation).

Ultimately I decided to keep the approach where the compiler
prefers coercing to a non-forwarding method over forwarding
to a method without coercing. I think that is reasonable
for the moment.

Nonetheless there is a bug in this area as shown by
delegate-coerce2.future.