Stop specifying the absolute path to "sphinx-build" when generating chpldocs
When generating the sphinx output we used to provide an absolute path to the
sphinx-build in chpl-venv, which made it impossible to use a locally installed
sphinx-build. This just switches it to call sphinx-build so that uses can build
the docs if they have all the requirements installed locally. Note that the
call to sphinx-build is preceded by a `export PATH=venvBinDir:$PATH`, so the
sphinx-build in the venv will still take precedence if it exists.