Fix bug where records don't recognize presence of super.init() call
Because records don't know what "super" means, the set up for detecting the
presence of "super.init()" calls on classes doesn't apply to them.  Add a check
that recognizes how the AST would look on a record, so I stop failing to
recognize when Phase 1 occurs for a record's initializer.