KNL Locale Model (#5555)
[Lots of much-appreciated review from @gbtitus , @ronawho , @mppf ]

This change implements a locale model for KNL. It adds four member functions for public use. These are implemented in all locale models, so that users can write portable code and yet use the advantages of specific types of machines. Locale models that have only one type of memory stub these functions out to point to that memory.

locale.defaultMemory() can get you back to normal memory if you are nested inside an on statement that uses a different kind of memory. On KNL, this is normal off-chip DDR memory.

locale.largeMemory() for whatever type of memory is the largest. On KNL, this is DDR memory.

locale.lowLatencyMemory() for low latency memory. On KNL, this is DDR memory.

locale.highBandwidthMemory() for high-bandwidth memory. On KNL, this is the on-chip MCDRAM.

The KNL NUMA domains that contain both compute cores and access to off-chip memory are the normal sublocales. For example, in KNL's SNC4 configuration, which has four compute-capable NUMA domains, these are numbered 0-3.

Internally, the KNL NUMA domains that contain only on-package MCDRAM are numbered consecutively above the compute sublocales. For example, in the SNC4 configuration, these are numbered 4-7. Sublocale 4's memory is closest to sublocale zero's cores, 5 is closest to 1, etc. If sublocale 4 is to be used, it means to compute in sublocale zero and store data in its closest MCDRAM bank. Users will normally access these sublocales with the member functions rather than having to know their number.

If any sublocale can be used, and data is to be stored in normal off-package memory, the sublocale ID is c_sublocid_any (which is negative 2). If any sublocale can be used, and data is to be stored in on-chip MCDRAM, the sublocale ID is one greater than the highest MCDRAM bank (on SNC4, this is sublocale 8).

As an example, to specify locale zero, sublocale 2, MCDRAM, you can use the following.

Locales(0).getChild(2).streamingMemory()

The following specifies any sublocale, and MCDRAM.

Locales(0).streamingMemory()

You can also specify the same locale as an existing variable, and still control which type of memory is used.

var x;
on x.locale.streamingMemory() { ... }

If the KNL node you are using is booted in cache mode (where MCDRAM is not accessible except as a level 3 cache), the streamingMemory() function just gives you normal DDR memory.

The KNL locale model is chosen by setting CHPL_LOCALE_MODEL=knl .
The following libraries are required.
hwloc
memkind
libnuma

PR #5566 (already merged) is a prerequisite for this PR.

Issue #5569 discusses items that this PR uncovered that should be addressed in the future.