make paratest.server work within salloc on Cray systems (#5600)
The paratest.server script is designed to ssh into the worker nodes when inside an salloc. This works for a cluster, but on a Cray X-series system, it tries to ssh into the compute nodes. Turning off the slurm detection when on a Cray X-series machine allows it to work under salloc there. This is a two-line change to do that.