Overlay nightly graphs onto releaseOverRelease graphs
When looking at the releaseOverRelease graphs we often open up the nightly
graphs at the same time so that we can see if regressions are within the
nightly noise or if it's something real to investigate. This allows us to view
releaseOverRelease and the nightly results on the same graph to simplify
releaseOverRelease performance triage.