Remove bashism from clean/cleanall/clobber targets
I've been working on an Ubuntu 17.04 system and there, Makefile commands
run with /bin/sh. A result of that is that an if statement like

  if [ -a doc/Makefile ]

results in a syntax error, since -a is a bash-ism.

This commit simply changes it to use -e which as far as I can tell does
exactly the same thing (does the file exist?) and is compatible with
/bin/sh.