Rename name field from NumaDomain and add tests
The name field is confusingly named because the parent class locale
already has a parentheses-less method named `proc name`.

The recent PR #5561 changed function resolution in a way that revealed
this issue.

I'm not sure the language should support for such overloading, but even
if it does, I don't think we should use it in NumaDomain because we'll
just confuse ourselves.

So, this PR simply renames the name field to ndName.

It adds two futures. One future tracks the exact issue we were running
into with NumaDomain and is related to const-checking in a constructor.
The second future asks about how ref-pairs should be resolved when one of
the elements of the ref-pair would normally have lower resolution
precedence (e.g. requires coercion, in a different scope, etc).