Nightly build fails if Chapel test elapsed time is negative
Once a month or so, a nightly Jenkins correctness test build fails
completely, because the script that reads the build log and writes
the JUnit result XML file dies. For some reason, a negative value
of "elapsed time" gets emitted into the log by something upstream
(probably sub_test or start_test, not sure), and that makes this
script throw an exception. The entire test build, including all
the test results, is lost.

There is reason to believe these negative elapsed times are due to
some kind of system clock skew on the Jenkins build slave, which is
always a VM when this happens.

Whatever is going on, there's no reason we have to lose the entire
build.

This change fixes util/test/convert_start_test_log_to_junit_xml.py
to be able to handle negative values of "elapsed time" without
throwing an exception.