Fix ref-pair bug
PR #5561 introduced a bug. When working with ref-pairs, resolveNormalCall
includes a series of calls to isBetterMatch that select ref/value
versions in cases where one is a better match than the other.

PR #5561 extended the ref-pair logic to treat const ref and value return
intents differently, but in doing so, it didn't extend this logic to call
isBetterMatch to choose between ref and non-ref versions.