Rename name field from MemoryLocale and NumaDomain in KNL locale model (#5628)
This is the KNL analog to the NUMA fix by @mppf in #5616 . It renames the "name" field in MemoryLocale and NumaDomain in the KNL locale model.

Many tests still fail while trying to compile module code in both the NUMA and KNL locale models on master. This PR just makes sure that no more tests fail in KNL than fail in NUMA.

Trivial, not reviewed.